+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Feature: automatically detect RSS/Atom feed from URL
+- Feature: automatically detect system color mode
 
 ### Changed